Testv2
Added while to repeat program while it is on. During the program the cart status is read in a if statement to determine which status option will be displayed to user. This operation is repeated by a for loop that once finished is followed by a sleep timer, which when complete launches the process over again. In the future the sleep timer section of the code should be changed to be updated by a pot rather than manually changing the value